Refactor: move Room and Marine into a Model folder
